requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
pydantic==2.11.4
pydantic-settings==2.9.1
python-multipart==0.0.20
google-genai==1.13.0
openai==1.76.2
